requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
colorama>=0.4.6
typing-extensions>=4.8.0
tenacity>=8.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0